requirements.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
awscli-local==0.14
backcall==0.2.0
bleach==3.3.0
boto3==1.17.83
botocore==1.20.83
branca==0.4.1
Brotli==1.0.9
Cerberus==1.3.2
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.1
entrypoints==0.3
enum34==1.1.10
Eve==1.1.5
Eve-Swagger==0.1.4
Events==0.3
falcon==2.0.0
falcon-jsonify==1.2
filelock==3.0.12
Fiona==1.8.17
Flask==2.0.1
Flask-Compress==1.9.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
future==0.18.2
geographiclib==1.50
geojson==2.5.0
geopandas==0.8.1
geopy==2.0.0
graphviz==0.17
greenlet==1.0.0
gunicorn==20.1.0
haversine==2.3.0
idna==2.10
ipykernel==5.5.3
ipyleaflet==0.13.3
ipympl==0.5.8
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlab-widgets==1.0.0
jupyterthemes==0.20.0
kiwisolver==1.2.0
lesscpy==0.14.0
localstack-client==1.20
lxml==4.5.2
MarkupSafe==2.0.1
matplotlib==3.3.2
mistune==0.8.4
mongoengine==0.22.1
munch==2.5.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.14.3
ply==3.11
polyline==1.4.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psutil==5.8.0
psycopg2==2.8.6
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pymongo==3.11.3
pyparsing==2.4.7
pyproj==2.6.1.post1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-mimeparse==1.6.0
python-statemachine==0.8.0
pytz==2021.1
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.24.0
retrying==1.3.3
rsome==0.1.3
s3transfer==0.4.2
Send2Trash==1.5.0
Shapely==1.7.1
sidecar==0.4.0
simplejson==3.17.3
six==1.15.0
terminado==0.9.4
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
traittypes==0.2.1
urllib3==1.25.10
virtualenv==20.0.31
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
